create_controlled_network_coupling function
